@@ -15,7 +15,7 @@
   "license": "Apache-2.0",
   "devDependencies": {
     "@types/node": "^10.17.0",
-    "typescript": "~3.7.2"
+    "typescript": "^3.8"
   },
   "dependencies": {
     "@aws-cdk/aws-certificatemanager": "*",
