@@ -1,5 +1,5 @@
 #!/usr/bin/env node
-import cdk = require('@aws-cdk/core');
+import * as cdk from '@aws-cdk/core';
 import { StaticSite } from './static-site';
 
 /**
@@ -9,7 +9,8 @@ import { StaticSite } from './static-site';
  * {
  *   "context": {
  *     "domain": "mystaticsite.com",
- *     "subdomain": "www"
+ *     "subdomain": "www",
+ *     "accountId": 1234567890,
  *   }
  * }
 **/
@@ -21,15 +22,27 @@ class MyStaticSiteStack extends cdk.Stack {
             domainName: this.node.tryGetContext('domain'),
             siteSubDomain: this.node.tryGetContext('subdomain'),
         });
-   }
+    }
 }
 
 const app = new cdk.App();
 
-new MyStaticSiteStack(app, 'MyStaticSite', { env: {
-    // Stack must be in us-east-1, because the ACM certificate for a
-    // global CloudFront distribution must be requested in us-east-1.
-    region: 'us-east-1'
-}});
+new MyStaticSiteStack(app, 'MyStaticSite', {
+    /**
+     * This is required for our use of hosted-zone lookup.
+     *
+     * Lookups do not work at all without an explicit environment
+     * specified; to use them, you must specify env.
+     * @see https://docs.aws.amazon.com/cdk/latest/guide/environments.html
+     */
+    env: {
+        account: app.node.tryGetContext('accountId'),
+        /**
+         * Stack must be in us-east-1, because the ACM certificate for a
+         * global CloudFront distribution must be requested in us-east-1.
+         */
+        region: 'us-east-1',
+    }
+});
 
 app.synth();
