@@ -1,12 +1,13 @@
 #!/usr/bin/env node
-import cloudfront = require('@aws-cdk/aws-cloudfront');
-import route53 = require('@aws-cdk/aws-route53');
-import s3 = require('@aws-cdk/aws-s3');
-import s3deploy = require('@aws-cdk/aws-s3-deployment');
-import acm = require('@aws-cdk/aws-certificatemanager');
-import cdk = require('@aws-cdk/core');
-import targets = require('@aws-cdk/aws-route53-targets/lib');
-import { Construct } from '@aws-cdk/core';
+import * as cdk from '@aws-cdk/core';
+import * as route53 from '@aws-cdk/aws-route53';
+import * as s3 from '@aws-cdk/aws-s3';
+import * as s3deploy from '@aws-cdk/aws-s3-deployment';
+import * as acm from '@aws-cdk/aws-certificatemanager';
+import * as targets from '@aws-cdk/aws-route53-targets';
+import * as cloudfront from '@aws-cdk/aws-cloudfront';
+import * as iam from '@aws-cdk/aws-iam';
+import { Construct, Stack } from '@aws-cdk/core';
 
 export interface StaticSiteProps {
     domainName: string;
@@ -20,11 +21,15 @@ export interface StaticSiteProps {
  * Route53 alias record, and ACM certificate.
  */
 export class StaticSite extends Construct {
-    constructor(parent: Construct, name: string, props: StaticSiteProps) {
+    constructor(parent: Stack, name: string, props: StaticSiteProps) {
         super(parent, name);
 
         const zone = route53.HostedZone.fromLookup(this, 'Zone', { domainName: props.domainName });
         const siteDomain = props.siteSubDomain + '.' + props.domainName;
+        const cloudfrontOAI = new cloudfront.OriginAccessIdentity(this, 'cloudfront-OAI', {
+            comment: `OAI for ${name}`
+        });
+
         new cdk.CfnOutput(this, 'Site', { value: 'https://' + siteDomain });
 
         // Content bucket
@@ -32,13 +37,28 @@ export class StaticSite extends Construct {
             bucketName: siteDomain,
             websiteIndexDocument: 'index.html',
             websiteErrorDocument: 'error.html',
-            publicReadAccess: true,
+            publicReadAccess: false,
+            blockPublicAccess: s3.BlockPublicAccess.BLOCK_ALL,
 
-            // The default removal policy is RETAIN, which means that cdk destroy will not attempt to delete
-            // the new bucket, and it will remain in your account until manually deleted. By setting the policy to
-            // DESTROY, cdk destroy will attempt to delete the bucket, but will error if the bucket is not empty.
+            /**
+             * The default removal policy is RETAIN, which means that cdk destroy will not attempt to delete
+             * the new bucket, and it will remain in your account until manually deleted. By setting the policy to
+             * DESTROY, cdk destroy will attempt to delete the bucket, but will error if the bucket is not empty.
+             */
             removalPolicy: cdk.RemovalPolicy.DESTROY, // NOT recommended for production code
+
+            /**
+             * For sample purposes only, if you create an S3 bucket then populate it, stack destruction fails.  This
+             * setting will enable full cleanup of the demo.
+             */
+            autoDeleteObjects: true, // NOT recommended for production code
         });
+        // Grant access to cloudfront
+        siteBucket.addToResourcePolicy(new iam.PolicyStatement({
+            actions: ['s3:GetObject'],
+            resources: [siteBucket.arnForObjects('*')],
+            principals: [new iam.CanonicalUserPrincipal(cloudfrontOAI.cloudFrontOriginAccessIdentityS3CanonicalUserId)]
+        }));
         new cdk.CfnOutput(this, 'Bucket', { value: siteBucket.bucketName });
 
         // TLS certificate
@@ -49,21 +69,36 @@ export class StaticSite extends Construct {
         }).certificateArn;
         new cdk.CfnOutput(this, 'Certificate', { value: certificateArn });
 
-        // CloudFront distribution that provides HTTPS
-        const distribution = new cloudfront.CloudFrontWebDistribution(this, 'SiteDistribution', {
-            aliasConfiguration: {
-                acmCertRef: certificateArn,
-                names: [ siteDomain ],
+        // Specifies you want viewers to use HTTPS & TLS v1.1 to request your objects
+        const viewerCertificate = cloudfront.ViewerCertificate.fromAcmCertificate({
+            certificateArn: certificateArn,
+            env: {
+                region: cdk.Aws.REGION,
+                account: cdk.Aws.ACCOUNT_ID
+            },
+            node: this.node,
+            stack: parent
+        },
+            {
                 sslMethod: cloudfront.SSLMethod.SNI,
                 securityPolicy: cloudfront.SecurityPolicyProtocol.TLS_V1_1_2016,
-            },
+                aliases: [siteDomain]
+            })
+
+        // CloudFront distribution
+        const distribution = new cloudfront.CloudFrontWebDistribution(this, 'SiteDistribution', {
+            viewerCertificate,
             originConfigs: [
                 {
-                    customOriginSource: {
-                        domainName: siteBucket.bucketWebsiteDomainName,
-                        originProtocolPolicy: cloudfront.OriginProtocolPolicy.HTTP_ONLY,
-                    },          
-                    behaviors : [ {isDefaultBehavior: true}],
+                    s3OriginSource: {
+                        s3BucketSource: siteBucket,
+                        originAccessIdentity: cloudfrontOAI
+                    },
+                    behaviors: [{
+                        isDefaultBehavior: true,
+                        compress: true,
+                        allowedMethods: cloudfront.CloudFrontAllowedMethods.GET_HEAD_OPTIONS,
+                    }],
                 }
             ]
         });
@@ -78,10 +113,10 @@ export class StaticSite extends Construct {
 
         // Deploy site contents to S3 bucket
         new s3deploy.BucketDeployment(this, 'DeployWithInvalidation', {
-            sources: [ s3deploy.Source.asset('./site-contents') ],
+            sources: [s3deploy.Source.asset('./site-contents')],
             destinationBucket: siteBucket,
             distribution,
             distributionPaths: ['/*'],
-          });
+        });
     }
 }
